# [PATCH] SemaphoreSlim, fail faster for timeout 0