@ludovic-henry 2016-09-14
LGTM. Is there anything blocking this PR from getting merged?