@marek-safar 2016-09-15
Yes, we'll include the file directly