@alexanderkyte 2018-04-12
@monojenkins build Linux x64 FullAOT